requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
torchaudio

openai>=1.12.0
sentence-transformers>=2.2.0
chromadb>=0.4.22
pypdf>=4.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
markdown>=3.5.0
gradio>=4.0.0
